pyproject.toml
[project]
name = "dev-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Rob Caduyac",email = "rmcaduyac1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.14"
dependencies = [
    "langchain (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.9,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "langgraph (>=0.3.18,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
